Rename AppDeployment extension to scenario-specific WinUI
The "AppDeployment" name was too generic: it over-promised a broad deployment
feature while the implementation only demonstrated the ITestHostLauncher hook.
Rename the consuming extension to Microsoft.Testing.Extensions.WinUI, anchoring it
to the concrete motivating scenario (#2784) instead.

- Microsoft.Testing.Extensions.WinUI: AddWinUIDeployment, WinUITestHostLauncher,
  WinUITestHostHandle (ProcessId => null), TestingPlatformBuilderHook (new GUID),
  build props, PublicAPI, PACKAGE.md; renamed in TestFx.slnx.
- The launcher is framed for WinUI: it implements the unpackaged deploy-and-launch
  path (stage the loose layout, launch the deployed app) and documents the packaged
  AUMID-activation branch as clearly-marked follow-up.
- Acceptance test renamed to WinUIDeploymentTests and gated to Windows; still proves
  end-to-end deploy + PID-less launch against the packed package.
- RFC non-goal updated: a reference WinUI consumer now exists (unpackaged path);
  packaged AUMID activation remains a separate follow-up.
