--- src/Platform/Microsoft.Testing.Extensions.AppDeployment/PublicAPI/PublicAPI.Unshipped.txt
@@ -1,5 +1,0 @@
-#nullable enable
-Microsoft.Testing.Extensions.AppDeployment.TestingPlatformBuilderHook
-Microsoft.Testing.Extensions.AppDeploymentExtensions
-static Microsoft.Testing.Extensions.AppDeployment.TestingPlatformBuilderHook.AddExtensions(Microsoft.Testing.Platform.Builder.ITestApplicationBuilder! testApplicationBuilder, string![]! _) -> void
-static Microsoft.Testing.Extensions.AppDeploymentExtensions.AddAppDeployment(this Microsoft.Testing.Platform.Builder.ITestApplicationBuilder! builder) -> void
