--- src/Platform/Microsoft.Testing.Extensions.AppDeployment/buildMultiTargeting/Microsoft.Testing.Extensions.AppDeployment.props
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Project>
-  <ItemGroup>
-    <TestingPlatformBuilderHook Include="4D48AEBF-3531-4EBF-AAF0-435D56A5FD00">
-      <DisplayName>Microsoft.Testing.Extensions.AppDeployment</DisplayName>
-      <TypeFullName>Microsoft.Testing.Extensions.AppDeployment.TestingPlatformBuilderHook</TypeFullName>
-    </TestingPlatformBuilderHook>
-  </ItemGroup>
-</Project>
