--- src/Platform/Microsoft.Testing.Extensions.AppDeployment/buildTransitive/Microsoft.Testing.Extensions.AppDeployment.props
@@ -1,3 +1,0 @@
-<Project>
-  <Import Project="$(MSBuildThisFileDirectory)..\..\buildMultiTargeting\Microsoft.Testing.Extensions.AppDeployment.props" />
-</Project>
