+++ src/Platform/Microsoft.Testing.Extensions.WinUI/BannedSymbols.txt
