+++ src/Platform/Microsoft.Testing.Extensions.WinUI/PublicAPI/PublicAPI.Unshipped.txt
@@ -1,0 +1,5 @@
+#nullable enable
+Microsoft.Testing.Extensions.WinUI.TestingPlatformBuilderHook
+Microsoft.Testing.Extensions.WinUIExtensions
+static Microsoft.Testing.Extensions.WinUI.TestingPlatformBuilderHook.AddExtensions(Microsoft.Testing.Platform.Builder.ITestApplicationBuilder! testApplicationBuilder, string![]! _) -> void
+static Microsoft.Testing.Extensions.WinUIExtensions.AddWinUIDeployment(this Microsoft.Testing.Platform.Builder.ITestApplicationBuilder! builder) -> void
