+++ src/Platform/Microsoft.Testing.Extensions.WinUI/buildMultiTargeting/Microsoft.Testing.Extensions.WinUI.props
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Project>
+  <ItemGroup>
+    <TestingPlatformBuilderHook Include="F7404E01-0F73-46B0-AC14-2E03993998BA">
+      <DisplayName>Microsoft.Testing.Extensions.WinUI</DisplayName>
+      <TypeFullName>Microsoft.Testing.Extensions.WinUI.TestingPlatformBuilderHook</TypeFullName>
+    </TestingPlatformBuilderHook>
+  </ItemGroup>
+</Project>
